test: separate mocks, add new abstract method

+++ src/applications/use_case/_test/delete-reply-use-case.test.ts
@@ -1,32 +1,11 @@
 import { describe, expect, it } from '@jest/globals';
 
-import { CommentRepository } from '@/domains/comments/comment-repository';
-import { ReplyRepository } from '@/domains/replies/reply-repository';
-import { ThreadRepository } from '@/domains/threads/thread-repository';
-
+import { MockCommentRepository } from '@/applications/use_case/_mocks/mock-comment-repository';
+import { MockReplyRepository } from '@/applications/use_case/_mocks/mock-reply-repository';
+import { MockThreadRepository } from '@/applications/use_case/_mocks/mock-thread-repository';
 import { DeleteReplyUseCase } from '@/applications/use_case/delete-reply-use-case';
 
 describe('DeleteReplyUseCase', () => {
-  class MockThreadRepository extends ThreadRepository {
-    verifyAvailableThread = jest.fn().mockResolvedValue(true);
-    addThread = jest.fn();
-    getThreadById = jest.fn();
-  }
-
-  class MockCommentRepository extends CommentRepository {
-    addComment = jest.fn();
-    verifyAvailableComment = jest.fn().mockResolvedValue(true);
-    deleteComment = jest.fn();
-    verifyCommentOwner = jest.fn().mockResolvedValue(true);
-  }
-
-  class MockReplyRepository extends ReplyRepository {
-    addReply = jest.fn();
-    verifyAvailableReply = jest.fn().mockResolvedValue(true);
-    deleteReply = jest.fn();
-    verifyReplyOwner = jest.fn().mockResolvedValue(true);
-  }
-
   it('should orchestrating the delete reply action correctly', async () => {
     // Arrange
     const useCasePayload = {
@@ -39,6 +18,20 @@ describe('DeleteReplyUseCase', () => {
     const mockThreadRepository = new MockThreadRepository();
     const mockCommentRepository = new MockCommentRepository();
     const mockReplyRepository = new MockReplyRepository();
+
+    mockThreadRepository.verifyAvailableThread = jest
+      .fn()
+      .mockResolvedValue(true);
+    mockCommentRepository.verifyAvailableComment = jest
+      .fn()
+      .mockResolvedValue(true);
+    mockCommentRepository.verifyCommentOwner = jest
+      .fn()
+      .mockResolvedValue(true);
+    mockReplyRepository.verifyAvailableReply = jest
+      .fn()
+      .mockResolvedValue(true);
+    mockReplyRepository.verifyReplyOwner = jest.fn().mockResolvedValue(true);
 
     const deleteReplyUseCase = new DeleteReplyUseCase({
       threadRepository: mockThreadRepository,
@@ -111,6 +104,10 @@ describe('DeleteReplyUseCase', () => {
     const mockCommentRepository = new MockCommentRepository();
     const mockReplyRepository = new MockReplyRepository();
 
+    mockThreadRepository.verifyAvailableThread = jest
+      .fn()
+      .mockResolvedValue(true);
+
     mockCommentRepository.verifyAvailableComment = jest
       .fn()
       .mockResolvedValue(false);
@@ -141,6 +138,12 @@ describe('DeleteReplyUseCase', () => {
     const mockCommentRepository = new MockCommentRepository();
     const mockReplyRepository = new MockReplyRepository();
 
+    mockThreadRepository.verifyAvailableThread = jest
+      .fn()
+      .mockResolvedValue(true);
+    mockCommentRepository.verifyAvailableComment = jest
+      .fn()
+      .mockResolvedValue(true);
     mockReplyRepository.verifyAvailableReply = jest
       .fn()
       .mockResolvedValue(false);
@@ -171,6 +174,15 @@ describe('DeleteReplyUseCase', () => {
     const mockCommentRepository = new MockCommentRepository();
     const mockReplyRepository = new MockReplyRepository();
 
+    mockThreadRepository.verifyAvailableThread = jest
+      .fn()
+      .mockResolvedValue(true);
+    mockCommentRepository.verifyAvailableComment = jest
+      .fn()
+      .mockResolvedValue(true);
+    mockReplyRepository.verifyAvailableReply = jest
+      .fn()
+      .mockResolvedValue(true);
     mockReplyRepository.verifyReplyOwner = jest.fn().mockResolvedValue(false);
 
     const deleteReplyUseCase = new DeleteReplyUseCase({
